Refactor modules, introduce sops-nix and NVF
Modules are now split a bit more nicely and I replace the alacritty
themes directory with a flake input.

For now, secrets are just generated to some file and then imported (the
hyrpland config) or picked up automatically (udev rule), so I didn't
have to worry about templates.

I also added my waybar config, which was missing before.

The biggest change is the use of NVF for configuring Neovim. I took the
plunge and decided to look at Nixvim and NVF and, being the hipster that
I am, I chose the newest one. Jokes aside, it's really nice to work
with. Right now the config is not complete, but it's pretty usable already.
I will keep the manual Lua config at least until I've migrated everything
useful in it to NVF.

Thanks a lot for the feedback and ideas, @rutrum!

Closes #2.